+++ packages/core/tests/legends/LegendItem.test.tsx
@@ -5,11 +5,15 @@ import { getNumberAttr } from '../utils'
 import { scaleCategorical, viewSeriesIndexesKeys } from './Legend.test'
 
 describe('LegendItem', () => {
-    const legendItemProps: Pick<LegendItemProps, 'position' | 'item' | 'label' | 'symbol'> = {
+    const legendItemProps: Pick<
+        LegendItemProps,
+        'position' | 'item' | 'label' | 'symbol' | 'disabledStyle'
+    > = {
         position: [0, 0],
         item: 'alpha',
         label: 'alpha',
         symbol: Circle,
+        disabledStyle: { opacity: 0.25 },
     }
 
     it('creates a legend item (variant right)', () => {
@@ -98,21 +102,21 @@ describe('LegendItem', () => {
             </Chart>
         )
         const item = screen.getByRole('legend-item')
-        expect(item.getAttribute('style')).toContain('opacity: 1')
+        expect(item.getAttribute('style') ?? '').not.toContain('opacity')
         // toggle to disabled state
         fireEvent.click(item)
         const item2 = await screen.findByRole('legend-item')
-        expect(item2.getAttribute('style')).not.toContain('opacity: 1')
+        expect(item2.getAttribute('style') ?? '').toContain('opacity: 0.25')
         const show = await screen.findByRole('show-disabled')
         expect(show.textContent).toContain('alpha')
         // toggle back to active state
         fireEvent.click(item)
         const item3 = await screen.findByRole('legend-item')
-        expect(item3.getAttribute('style')).toContain('opacity: 1')
+        expect(item3.getAttribute('style') ?? '').not.toContain('opacity')
         // toggle again to disabled state
         fireEvent.click(item)
         const item4 = await screen.findByRole('legend-item')
-        expect(item4.getAttribute('style')).not.toContain('opacity: 1')
+        expect(item4.getAttribute('style') ?? '').toContain('opacity: 0.25')
     })
 
     it('does not toggle when interactivity is turned off', async () => {
@@ -126,9 +130,9 @@ describe('LegendItem', () => {
             </Chart>
         )
         const item = screen.getByRole('legend-item')
-        expect(item.getAttribute('style')).toContain('opacity: 1')
+        expect(item.getAttribute('style') ?? '').not.toContain('opacity')
         fireEvent.click(item)
         const item2 = await screen.findByRole('legend-item')
-        expect(item2.getAttribute('style')).toContain('opacity: 1')
+        expect(item2.getAttribute('style') ?? '').not.toContain('opacity')
     })
 })